Prepare to deploy

+++ server/common/src/main/java/com/sawwere/titlecounter/common/dto/film/FilmEntryRequestDto.java
@@ -24,7 +24,7 @@ public class FilmEntryRequestDto {
 
     @Min(1)
     @Max(10)
-    private Long score;
+    private Integer score;
 
     @NotBlank
     private String status;
